+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+- `measurementUnit` text and `unitMultiplier` logic at `product-selling-price` and `product-list-price`.
 
 ## [1.15.0] - 2021-03-09
 ### Added
-- `percentageStyle` on `product-price-savings`
+- `percentageStyle` on `product-price-savings`.
 
 ## [1.14.0] - 2021-03-08
 
@@ -21,12 +24,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.13.1] - 2021-03-08
 
 ### Fixed
-- Updated documentation (removed link to internal doc)
+- Updated documentation (removed link to internal doc).
 
 ## [1.13.0] - 2021-02-04
 
 ### Added
-- `taxValue` on `product-selling-price` and `product-list-price`
+- `taxValue` on `product-selling-price` and `product-list-price`.
 
 ## [1.12.0] - 2021-01-06
 ### Added